Stop packaging Libtool `.la` files.
Pursuant to https://github.com/conda-forge/staged-recipes/issues/673 and
https://github.com/conda-forge/libxcb-feedstock/pull/9, we're going to remove
these files from our packages. Eventually conda-build will do this
automatically for us, but for now we take care of it manually. We need to
delete the files both before the build (because inconsistent complements of
`.la` files can break the build) and after (because we may have installed new
ones).